Move `disable` to `NotificationsNamespace`
